requirements.txt
django==3.1.4
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-simple-history==2.12.0
django-extensions==3.1.0
django-cacheops==5.1
django-redis==4.12.1
celery==4.4.6
pyjwt==1.7.1
dateutils==0.6.12
requests==2.26.0
boto3==1.20.24
openpyxl==3.0.9
django-filter==21.1
drf-yasg==1.20.0
cryptography==3.4.7
